Revamp categories handling.
Move categories list in `static_content.cpp`
We use two "string" to handle each category. The first one is the
tag to use to do the opds request. The second one is the one displayed to
the user. The later is translatable.
The list is taken from issue kiwix/kiwix-tools#317

Fix #193